+++ shared/src/postgres/transaction.rs
@@ -8,6 +8,9 @@ use std::{
 use thiserror::Error;
 use tracing::error;
 
+static TRANSACTION_CALL_ONCE_ON_PRECOMMIT_CACHE: Lazy<Arc<Mutex<HashSet<String>>>> =
+    Lazy::new(|| Arc::new(Mutex::new(HashSet::new())));
+
 static TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE: Lazy<Arc<Mutex<HashSet<String>>>> =
     Lazy::new(|| Arc::new(Mutex::new(HashSet::new())));
 
@@ -18,7 +21,7 @@ pub struct Transaction {}
 
 impl Transaction {
     pub fn needs_commit(id: &str) -> Result<bool, TransactionError> {
-        let cache = TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE.lock()?;
+        let cache = TRANSACTION_CALL_ONCE_ON_PRECOMMIT_CACHE.lock()?;
         Ok(cache.contains(id))
     }
 
@@ -27,12 +30,41 @@ impl Transaction {
         F: FnOnce() + Send + UnwindSafe + RefUnwindSafe + 'static,
     {
         // Clone the cache here for use inside the closure.
-        let cache_clone = TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE.clone();
-        let mut cache = TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE.lock()?;
+        let cache_clone = TRANSACTION_CALL_ONCE_ON_PRECOMMIT_CACHE.clone();
+        let mut cache = TRANSACTION_CALL_ONCE_ON_PRECOMMIT_CACHE.lock()?;
 
         if !cache.contains(id) {
             // Now using `cache_clone` inside the closure.
             register_xact_callback(PgXactCallbackEvent::PreCommit, move || {
+                // Clear the cache so callbacks can be registered on next transaction.
+                match cache_clone.lock() {
+                    Ok(mut cache) => cache.clear(),
+                    Err(err) => error!(
+                        "could not acquire lock in register transaction precommit callback: {err:?}"
+                    ),
+                }
+
+                // Actually call the callback.
+                callback();
+            });
+
+            cache.insert(id.into());
+        }
+
+        Ok(())
+    }
+
+    pub fn call_once_on_commit<F>(id: &str, callback: F) -> Result<(), TransactionError>
+    where
+        F: FnOnce() + Send + UnwindSafe + RefUnwindSafe + 'static,
+    {
+        // Clone the cache here for use inside the closure.
+        let cache_clone = TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE.clone();
+
+        let mut cache = TRANSACTION_CALL_ONCE_ON_COMMIT_CACHE.lock()?;
+        if !cache.contains(id) {
+            // Now using `cache_clone` inside the closure.
+            register_xact_callback(PgXactCallbackEvent::Commit, move || {
                 // Clear the cache so callbacks can be registered on next transaction.
                 match cache_clone.lock() {
                     Ok(mut cache) => cache.clear(),
